test: move promoted widget lookup to node reference
Addresses review feedback:
https://github.com/Comfy-Org/ComfyUI_frontend/pull/17516#discussion_r3995540431
